requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
Pillow
pydantic
python-multipart
